Outputs clean rh_corrected_nephelometer measures for analysis.

Outputs clean rh_corrected_nephelometer measures for analysis.

cleaningMeasures(micropemObject, hepaStart = FALSE, hepaEnd = FALSE)

Arguments

micropemObject

the MicroPEM object

hepaStart

Boolean indicating whether there were measurements with HEPA filters at the beginning.

hepaEnd

Boolean indicating whether there were measurements with HEPA filters at the end.

Value

A MicroPEM object.

Details

rh_corrected_nephelometer values are set to NA if they are negative or if the RH at the same time is higher than 90%. rh_corrected_nephelometer values are also corrected for the HEPA zeroings (start and end, if there were done): if a stable period longer than 3 minutes can be identified for the HEPA period, using the changepoint cpt.mean function, there is a zero value. There can be no zero values, only one (beginning or end) or two. If there is only one zero value, it is substracted from all rh_corrected_nephelometer values. If there are two, a linear interpolation is done between the two values and the resulting vector is substracted from the rh_corrected_nephelometer values.

Examples

data(micropemChai)
cleanMP <- cleaningMeasures(micropemChai)
cleanMP$summary()

#> # A tibble: 7 × 7
#>                     measure no._of_not_missing_values median       mean minimum
#>                       <chr>                     <int>  <dbl>      <dbl>   <dbl>
#> 1 rh_corrected_nephelometer                      8634  49.00 49.3745657   45.00
#> 2                      temp                      2878  84.50 84.6830438   82.30
#> 3                        rh                      8634  54.60 55.0061733   46.20
#> 4                   battery                      1464   4.10  4.0872268    3.90
#> 5               inlet_press                      2878   0.11  0.1111015    0.10
#> 6             orifice_press                      2878   0.15  0.1505455    0.14
#> 7                      flow                      2878   0.77  0.7703023    0.77
#> # ... with 2 more variables: maximum <dbl>, variance <dbl>